requirements.txt
opencv-python~=4.11.0.86
numpy~=1.26.4
pillow~=11.1.0
pyside6~=6.10.1